final working version
Converted function in function form to object oriented form. Working with Poisson, Negetive binomial and geometric zero inflated models. EM estimate of starting values for ZINB not available since NB regression function with weights as parameters and doing joint estimation of scale and mu is not available in python yet.